Fix settings module imports
