+++ crates/hydebar-core/src/modules/settings.rs
@@ -1,41 +1,7 @@
-use self::{
-    audio::AudioMessage, bluetooth::BluetoothMessage, network::NetworkMessage, power::PowerMessage,
-};
-use super::{Module, ModuleError, OnModulePress};
-use crate::{
-    ModuleContext, ModuleEventSender, app,
-    components::icons::{Icons, icon},
-    config::{Position, SettingsModuleConfig},
-    event_bus::ModuleEvent,
-    menu::MenuType,
-    modules::settings::power::power_menu,
-    outputs::Outputs,
-    password_dialog,
-    position_button::ButtonUIRef,
-    services::{
-        ReadOnlyService, Service, ServiceEvent, ServiceEventPublisher,
-        audio::{AudioCommand, AudioService},
-        bluetooth::{BluetoothCommand, BluetoothService, BluetoothState},
-        brightness::{BrightnessCommand, BrightnessService},
-        idle_inhibitor::IdleInhibitorManager,
-        network::{NetworkCommand, NetworkEvent, NetworkService},
-        upower::{PowerProfileCommand, UPowerService},
-    },
-    style::{
-        quick_settings_button_style, quick_settings_submenu_button_style, settings_button_style,
-    },
-};
-use brightness::BrightnessMessage;
-use iced::{
-    Alignment, Background, Border, Element, Length, Padding, Task, Theme,
-    alignment::{Horizontal, Vertical},
-    widget::{Column, Row, Space, button, column, container, horizontal_space, row, text},
-    window::Id,
-};
-use log::{info, warn};
-use std::future::{Ready, ready};
-use tokio::{runtime::Handle, task::JoinHandle};
-use upower::UPowerMessage;
+mod commands;
+mod event_forwarders;
+mod state;
+mod view;
 
 pub mod audio;
 pub mod bluetooth;
@@ -44,1044 +10,11 @@ pub mod network;
 mod power;
 mod upower;
 
-pub struct Settings {
-    audio: Option<AudioService>,
-    pub brightness: Option<BrightnessService>,
-    network: Option<NetworkService>,
-    bluetooth: Option<BluetoothService>,
-    idle_inhibitor: Option<IdleInhibitorManager>,
-    pub sub_menu: Option<SubMenu>,
-    upower: Option<UPowerService>,
-    pub password_dialog: Option<(String, String)>,
-    sender: Option<ModuleEventSender<Message>>,
-    runtime: Option<Handle>,
-    tasks: Vec<JoinHandle<()>>,
-}
-
-impl Default for Settings {
-    fn default() -> Self {
-        let idle_inhibitor = match IdleInhibitorManager::new() {
-            Ok(manager) => Some(manager),
-            Err(err) => {
-                warn!("Failed to initialize idle inhibitor: {err}");
-                None
-            }
-        };
-
-        Settings {
-            audio: None,
-            brightness: None,
-            network: None,
-            bluetooth: None,
-            idle_inhibitor,
-            sub_menu: None,
-            upower: None,
-            password_dialog: None,
-            sender: None,
-            runtime: None,
-            tasks: Vec::new(),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum Message {
-    ToggleMenu(Id, ButtonUIRef),
-    UPower(UPowerMessage),
-    Network(NetworkMessage),
-    Bluetooth(BluetoothMessage),
-    Audio(AudioMessage),
-    Brightness(BrightnessMessage),
-    ToggleInhibitIdle,
-    Lock,
-    Power(PowerMessage),
-    ToggleSubMenu(SubMenu),
-    PasswordDialog(password_dialog::Message),
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum SubMenu {
-    Power,
-    Sinks,
-    Sources,
-    Wifi,
-    Vpn,
-    Bluetooth,
-}
-
-impl Settings {
-    fn runtime(&self) -> Option<Handle> {
-        self.runtime.as_ref().cloned()
-    }
-
-    fn sender(&self) -> Option<ModuleEventSender<Message>> {
-        self.sender.as_ref().cloned()
-    }
-
-    fn spawn_audio_command(&self, command: AudioCommand) {
-        if let (Some(handle), Some(sender), Some(service)) =
-            (self.runtime(), self.sender(), self.audio.clone())
-        {
-            handle.spawn(async move {
-                if let Some(event) = AudioService::run_command(service, command).await {
-                    if let Err(err) = sender.try_send(Message::Audio(AudioMessage::Event(event))) {
-                        warn!("failed to publish audio command event: {err}");
-                    }
-                }
-            });
-        } else {
-            warn!("audio command ignored because runtime, sender, or service is unavailable");
-        }
-    }
-
-    fn spawn_brightness_command(&self, command: BrightnessCommand) {
-        if let (Some(handle), Some(sender), Some(service)) =
-            (self.runtime(), self.sender(), self.brightness.clone())
-        {
-            handle.spawn(async move {
-                let event = BrightnessService::run_command(service, command).await;
-                if let Err(err) =
-                    sender.try_send(Message::Brightness(BrightnessMessage::Event(event)))
-                {
-                    warn!("failed to publish brightness command event: {err}");
-                }
-            });
-        } else {
-            warn!("brightness command ignored because runtime, sender, or service is unavailable");
-        }
-    }
-
-    fn spawn_network_command(&self, command: NetworkCommand) {
-        if let (Some(handle), Some(sender), Some(service)) =
-            (self.runtime(), self.sender(), self.network.clone())
-        {
-            handle.spawn(async move {
-                let event = NetworkService::run_command(service, command).await;
-                if let Err(err) = sender.try_send(Message::Network(NetworkMessage::Event(event))) {
-                    warn!("failed to publish network command event: {err}");
-                }
-            });
-        } else {
-            warn!("network command ignored because runtime, sender, or service is unavailable");
-        }
-    }
-
-    fn spawn_bluetooth_command(&self, command: BluetoothCommand) {
-        if let (Some(handle), Some(sender), Some(service)) =
-            (self.runtime(), self.sender(), self.bluetooth.clone())
-        {
-            handle.spawn(async move {
-                if let Some(event) = BluetoothService::run_command(service, command).await {
-                    if let Err(err) =
-                        sender.try_send(Message::Bluetooth(BluetoothMessage::Event(event)))
-                    {
-                        warn!("failed to publish bluetooth command event: {err}");
-                    }
-                }
-            });
-        } else {
-            warn!("bluetooth command ignored because runtime, sender, or service is unavailable");
-        }
-    }
-
-    fn spawn_upower_command(&self, command: PowerProfileCommand) {
-        if let (Some(handle), Some(sender), Some(service)) =
-            (self.runtime(), self.sender(), self.upower.clone())
-        {
-            handle.spawn(async move {
-                let event = UPowerService::run_command(service, command).await;
-                if let Err(err) = sender.try_send(Message::UPower(UPowerMessage::Event(event))) {
-                    warn!("failed to publish upower command event: {err}");
-                }
-            });
-        } else {
-            warn!("upower command ignored because runtime, sender, or service is unavailable");
-        }
-    }
-
-    pub fn update(
-        &mut self,
-        message: Message,
-        config: &SettingsModuleConfig,
-        outputs: &mut Outputs,
-        main_config: &crate::config::Config,
-    ) -> Task<crate::app::Message> {
-        match message {
-            Message::ToggleMenu(id, button_ui_ref) => {
-                self.sub_menu = None;
-                self.password_dialog = None;
-                outputs.toggle_menu(id, MenuType::Settings, button_ui_ref, main_config)
-            }
-            Message::Audio(msg) => match msg {
-                AudioMessage::Event(event) => match event {
-                    ServiceEvent::Init(service) => {
-                        self.audio = Some(service);
-                        Task::none()
-                    }
-                    ServiceEvent::Update(data) => {
-                        if let Some(audio) = self.audio.as_mut() {
-                            audio.update(data);
-
-                            if self.sub_menu == Some(SubMenu::Sinks) && audio.sinks.len() < 2 {
-                                self.sub_menu = None;
-                            }
-
-                            if self.sub_menu == Some(SubMenu::Sources) && audio.sources.len() < 2 {
-                                self.sub_menu = None;
-                            }
-                        }
-                        Task::none()
-                    }
-                    ServiceEvent::Error(_) => Task::none(),
-                },
-                AudioMessage::ToggleSinkMute => {
-                    self.spawn_audio_command(AudioCommand::ToggleSinkMute);
-                    Task::none()
-                }
-                AudioMessage::SinkVolumeChanged(value) => {
-                    self.spawn_audio_command(AudioCommand::SinkVolume(value));
-                    Task::none()
-                }
-                AudioMessage::DefaultSinkChanged(name, port) => {
-                    self.spawn_audio_command(AudioCommand::DefaultSink(name, port));
-                    Task::none()
-                }
-                AudioMessage::ToggleSourceMute => {
-                    self.spawn_audio_command(AudioCommand::ToggleSourceMute);
-                    Task::none()
-                }
-                AudioMessage::SourceVolumeChanged(value) => {
-                    self.spawn_audio_command(AudioCommand::SourceVolume(value));
-                    Task::none()
-                }
-                AudioMessage::DefaultSourceChanged(name, port) => {
-                    self.spawn_audio_command(AudioCommand::DefaultSource(name, port));
-                    Task::none()
-                }
-                AudioMessage::SinksMore(id) => {
-                    if let Some(cmd) = &config.audio_sinks_more_cmd {
-                        crate::utils::launcher::execute_command(cmd.to_string());
-                        outputs.close_menu(id, main_config)
-                    } else {
-                        Task::none()
-                    }
-                }
-                AudioMessage::SourcesMore(id) => {
-                    if let Some(cmd) = &config.audio_sources_more_cmd {
-                        crate::utils::launcher::execute_command(cmd.to_string());
-                        outputs.close_menu(id, main_config)
-                    } else {
-                        Task::none()
-                    }
-                }
-            },
-            Message::UPower(msg) => match msg {
-                UPowerMessage::Event(event) => match event {
-                    ServiceEvent::Init(service) => {
-                        self.upower = Some(service);
-                        Task::none()
-                    }
-                    ServiceEvent::Update(data) => {
-                        if let Some(upower) = self.upower.as_mut() {
-                            upower.update(data);
-                        }
-                        Task::none()
-                    }
-                    ServiceEvent::Error(_) => Task::none(),
-                },
-                UPowerMessage::TogglePowerProfile => {
-                    self.spawn_upower_command(PowerProfileCommand::Toggle);
-                    Task::none()
-                }
-            },
-            Message::Network(msg) => match msg {
-                NetworkMessage::Event(event) => match event {
-                    ServiceEvent::Init(service) => {
-                        self.network = Some(service);
-                        Task::none()
-                    }
-                    ServiceEvent::Update(NetworkEvent::RequestPasswordForSSID(ssid)) => {
-                        self.password_dialog = Some((ssid, "".to_string()));
-                        Task::none()
-                    }
-                    ServiceEvent::Update(data) => {
-                        if let Some(network) = self.network.as_mut() {
-                            network.update(data);
-                        }
-                        Task::none()
-                    }
-                    _ => Task::none(),
-                },
-                NetworkMessage::ToggleAirplaneMode => {
-                    if self.sub_menu == Some(SubMenu::Wifi) {
-                        self.sub_menu = None;
-                    }
-
-                    self.spawn_network_command(NetworkCommand::ToggleAirplaneMode);
-                    Task::none()
-                }
-                NetworkMessage::ToggleWiFi => {
-                    if self.sub_menu == Some(SubMenu::Wifi) {
-                        self.sub_menu = None;
-                    }
-
-                    self.spawn_network_command(NetworkCommand::ToggleWiFi);
-                    Task::none()
-                }
-                NetworkMessage::SelectAccessPoint(ac) => {
-                    self.spawn_network_command(NetworkCommand::SelectAccessPoint((ac, None)));
-                    Task::none()
-                }
-                NetworkMessage::RequestWiFiPassword(id, ssid) => {
-                    info!("Requesting password for {ssid}");
-                    self.password_dialog = Some((ssid, "".to_string()));
-                    outputs.request_keyboard(id, main_config.menu_keyboard_focus)
-                }
-                NetworkMessage::ScanNearByWiFi => {
-                    self.spawn_network_command(NetworkCommand::ScanNearByWiFi);
-                    Task::none()
-                }
-                NetworkMessage::WiFiMore(id) => {
-                    if let Some(cmd) = &config.wifi_more_cmd {
-                        crate::utils::launcher::execute_command(cmd.to_string());
-                        outputs.close_menu(id, main_config)
-                    } else {
-                        Task::none()
-                    }
-                }
-                NetworkMessage::VpnMore(id) => {
-                    if let Some(cmd) = &config.vpn_more_cmd {
-                        crate::utils::launcher::execute_command(cmd.to_string());
-                        outputs.close_menu(id, main_config)
-                    } else {
-                        Task::none()
-                    }
-                }
-                NetworkMessage::ToggleVpn(vpn) => {
-                    self.spawn_network_command(NetworkCommand::ToggleVpn(vpn));
-                    Task::none()
-                }
-            },
-            Message::Bluetooth(msg) => match msg {
-                BluetoothMessage::Event(event) => match event {
-                    ServiceEvent::Init(service) => {
-                        self.bluetooth = Some(service);
-                        Task::none()
-                    }
-                    ServiceEvent::Update(data) => {
-                        if let Some(bluetooth) = self.bluetooth.as_mut() {
-                            bluetooth.update(data);
-                        }
-                        Task::none()
-                    }
-                    _ => Task::none(),
-                },
-                BluetoothMessage::Toggle => match self.bluetooth.as_mut() {
-                    Some(bluetooth) => {
-                        if self.sub_menu == Some(SubMenu::Bluetooth) {
-                            self.sub_menu = None;
-                        }
-
-                        self.spawn_bluetooth_command(BluetoothCommand::Toggle);
-                        Task::none()
-                    }
-                    _ => Task::none(),
-                },
-                BluetoothMessage::More(id) => {
-                    if let Some(cmd) = &config.bluetooth_more_cmd {
-                        crate::utils::launcher::execute_command(cmd.to_string());
-                        outputs.close_menu(id, main_config)
-                    } else {
-                        Task::none()
-                    }
-                }
-            },
-            Message::Brightness(msg) => match msg {
-                BrightnessMessage::Event(event) => match event {
-                    ServiceEvent::Init(service) => {
-                        self.brightness = Some(service);
-                        Task::none()
-                    }
-                    ServiceEvent::Update(data) => {
-                        if let Some(brightness) = self.brightness.as_mut() {
-                            brightness.update(data);
-                        }
-                        Task::none()
-                    }
-                    _ => Task::none(),
-                },
-                BrightnessMessage::Change(value) => {
-                    self.spawn_brightness_command(BrightnessCommand::Set(value));
-                    Task::none()
-                }
-            },
-            Message::ToggleSubMenu(menu_type) => {
-                if self.sub_menu == Some(menu_type) {
-                    self.sub_menu.take();
-                } else {
-                    self.sub_menu.replace(menu_type);
-
-                    if menu_type == SubMenu::Wifi {
-                        self.spawn_network_command(NetworkCommand::ScanNearByWiFi);
-                    }
-                }
-
-                Task::none()
-            }
-            Message::ToggleInhibitIdle => {
-                if let Some(idle_inhibitor) = &mut self.idle_inhibitor {
-                    idle_inhibitor.toggle();
-                }
-                Task::none()
-            }
-            Message::Lock => {
-                if let Some(lock_cmd) = &config.lock_cmd {
-                    crate::utils::launcher::execute_command(lock_cmd.to_string());
-                }
-                Task::none()
-            }
-            Message::Power(msg) => {
-                msg.update();
-                Task::none()
-            }
-            Message::PasswordDialog(msg) => match msg {
-                password_dialog::Message::PasswordChanged(password) => {
-                    if let Some((_, current_password)) = &mut self.password_dialog {
-                        *current_password = password;
-                    }
-
-                    Task::none()
-                }
-                password_dialog::Message::DialogConfirmed(id) => {
-                    if let Some((ssid, password)) = self.password_dialog.take() {
-                        if let Some(network) = self.network.as_ref() {
-                            if let Some(access_point) = network
-                                .wireless_access_points
-                                .iter()
-                                .find(|ap| ap.ssid == ssid)
-                                .cloned()
-                            {
-                                self.spawn_network_command(NetworkCommand::SelectAccessPoint((
-                                    access_point,
-                                    Some(password.clone()),
-                                )));
-                            }
-                        }
-
-                        outputs.release_keyboard(id, main_config.menu_keyboard_focus)
-                    } else {
-                        outputs.release_keyboard(id, main_config.menu_keyboard_focus)
-                    }
-                }
-                password_dialog::Message::DialogCancelled(id) => {
-                    self.password_dialog = None;
-
-                    outputs.release_keyboard(id, main_config.menu_keyboard_focus)
-                }
-            },
-        }
-    }
-
-    pub fn menu_view(
-        &self,
-        id: Id,
-        config: &SettingsModuleConfig,
-        opacity: f32,
-        position: Position,
-    ) -> Element<Message> {
-        if let Some((ssid, current_password)) = &self.password_dialog {
-            password_dialog::view(id, ssid, current_password, opacity).map(Message::PasswordDialog)
-        } else {
-            let battery_data = self
-                .upower
-                .as_ref()
-                .and_then(|upower| upower.battery)
-                .map(|battery| battery.settings_indicator());
-            let right_buttons = Row::new()
-                .push_maybe(config.lock_cmd.as_ref().map(|_| {
-                    button(icon(Icons::Lock))
-                        .padding([8, 13])
-                        .on_press(Message::Lock)
-                        .style(settings_button_style(opacity))
-                }))
-                .push(
-                    button(icon(if self.sub_menu == Some(SubMenu::Power) {
-                        Icons::Close
-                    } else {
-                        Icons::Power
-                    }))
-                    .padding([8, 13])
-                    .on_press(Message::ToggleSubMenu(SubMenu::Power))
-                    .style(settings_button_style(opacity)),
-                )
-                .spacing(8);
-
-            let header = Row::new()
-                .push_maybe(battery_data)
-                .push(Space::with_width(Length::Fill))
-                .push(right_buttons)
-                .spacing(8)
-                .width(Length::Fill);
-
-            let (sink_slider, source_slider) = self
-                .audio
-                .as_ref()
-                .map(|a| a.audio_sliders(self.sub_menu, opacity))
-                .unwrap_or((None, None));
-
-            let wifi_setting_button = self.network.as_ref().and_then(|n| {
-                n.get_wifi_quick_setting_button(
-                    id,
-                    self.sub_menu,
-                    config.wifi_more_cmd.is_some(),
-                    opacity,
-                )
-            });
-            let quick_settings = quick_settings_section(
-                vec![
-                    wifi_setting_button,
-                    self.bluetooth
-                        .as_ref()
-                        .filter(|b| b.state != BluetoothState::Unavailable)
-                        .and_then(|b| {
-                            b.get_quick_setting_button(
-                                id,
-                                self.sub_menu,
-                                config.bluetooth_more_cmd.is_some(),
-                                opacity,
-                            )
-                        }),
-                    self.network.as_ref().and_then(|n| {
-                        n.get_vpn_quick_setting_button(
-                            id,
-                            self.sub_menu,
-                            config.vpn_more_cmd.is_some(),
-                            opacity,
-                        )
-                    }),
-                    self.network.as_ref().and_then(|n| {
-                        if config.remove_airplane_btn {
-                            None
-                        } else {
-                            Some(n.get_airplane_mode_quick_setting_button(opacity))
-                        }
-                    }),
-                    self.idle_inhibitor.as_ref().and_then(|i| {
-                        if config.remove_idle_btn {
-                            None
-                        } else {
-                            Some((
-                                quick_setting_button(
-                                    if i.is_inhibited() {
-                                        Icons::EyeOpened
-                                    } else {
-                                        Icons::EyeClosed
-                                    },
-                                    "Idle Inhibitor".to_string(),
-                                    None,
-                                    i.is_inhibited(),
-                                    Message::ToggleInhibitIdle,
-                                    None,
-                                    opacity,
-                                ),
-                                None,
-                            ))
-                        }
-                    }),
-                    self.upower
-                        .as_ref()
-                        .and_then(|u| u.power_profile.get_quick_setting_button(opacity)),
-                ]
-                .into_iter()
-                .flatten()
-                .collect::<Vec<_>>(),
-                opacity,
-            );
-
-            let (top_sink_slider, bottom_sink_slider) = match position {
-                Position::Top => (sink_slider, None),
-                Position::Bottom => (None, sink_slider),
-            };
-            let (top_source_slider, bottom_source_slider) = match position {
-                Position::Top => (source_slider, None),
-                Position::Bottom => (None, source_slider),
-            };
-
-            Column::new()
-                .push(header)
-                .push_maybe(
-                    self.sub_menu
-                        .filter(|menu_type| *menu_type == SubMenu::Power)
-                        .map(|_| {
-                            sub_menu_wrapper(
-                                power_menu(opacity, config).map(Message::Power),
-                                opacity,
-                            )
-                        }),
-                )
-                .push_maybe(top_sink_slider)
-                .push_maybe(
-                    self.sub_menu
-                        .filter(|menu_type| *menu_type == SubMenu::Sinks)
-                        .and_then(|_| {
-                            self.audio.as_ref().map(|a| {
-                                sub_menu_wrapper(
-                                    a.sinks_submenu(
-                                        id,
-                                        config.audio_sinks_more_cmd.is_some(),
-                                        opacity,
-                                    ),
-                                    opacity,
-                                )
-                            })
-                        }),
-                )
-                .push_maybe(bottom_sink_slider)
-                .push_maybe(top_source_slider)
-                .push_maybe(
-                    self.sub_menu
-                        .filter(|menu_type| *menu_type == SubMenu::Sources)
-                        .and_then(|_| {
-                            self.audio.as_ref().map(|a| {
-                                sub_menu_wrapper(
-                                    a.sources_submenu(
-                                        id,
-                                        config.audio_sources_more_cmd.is_some(),
-                                        opacity,
-                                    ),
-                                    opacity,
-                                )
-                            })
-                        }),
-                )
-                .push_maybe(bottom_source_slider)
-                .push_maybe(self.brightness.as_ref().map(|b| b.brightness_slider()))
-                .push(quick_settings)
-                .spacing(16)
-                .into()
-        }
-    }
-}
-
-impl Module for Settings {
-    type ViewData<'a> = ();
-    type RegistrationData<'a> = ();
-
-    fn register(
-        &mut self,
-        ctx: &ModuleContext,
-        _: Self::RegistrationData<'_>,
-    ) -> Result<(), ModuleError> {
-        for task in self.tasks.drain(..) {
-            task.abort();
-        }
-
-        let sender = ctx.module_sender(ModuleEvent::Settings);
-
-        let mut tasks = Vec::new();
-
-        let mut audio_publisher = AudioEventForwarder::new(sender.clone());
-        tasks.push(ctx.runtime_handle().spawn(async move {
-            AudioService::listen(&mut audio_publisher).await;
-        }));
-
-        let mut brightness_publisher = BrightnessEventForwarder::new(sender.clone());
-        tasks.push(ctx.runtime_handle().spawn(async move {
-            BrightnessService::listen(&mut brightness_publisher).await;
-        }));
-
-        let mut network_publisher = NetworkEventForwarder::new(sender.clone());
-        tasks.push(ctx.runtime_handle().spawn(async move {
-            NetworkService::listen(&mut network_publisher).await;
-        }));
-
-        let mut bluetooth_publisher = BluetoothEventForwarder::new(sender.clone());
-        tasks.push(ctx.runtime_handle().spawn(async move {
-            BluetoothService::listen(&mut bluetooth_publisher).await;
-        }));
-
-        let mut upower_publisher = UPowerEventForwarder::new(sender.clone());
-        tasks.push(ctx.runtime_handle().spawn(async move {
-            UPowerService::listen(&mut upower_publisher).await;
-        }));
-
-        self.sender = Some(sender);
-        self.runtime = Some(ctx.runtime_handle().clone());
-        self.tasks = tasks;
-
-        Ok(())
-    }
-
-    fn view(
-        &self,
-        _: Self::ViewData<'_>,
-    ) -> Option<(Element<app::Message>, Option<OnModulePress>)> {
-        Some((
-            Row::new()
-                .push_maybe(
-                    self.idle_inhibitor
-                        .as_ref()
-                        .filter(|i| i.is_inhibited())
-                        .map(|_| {
-                            container(icon(Icons::EyeOpened)).style(|theme: &Theme| {
-                                container::Style {
-                                    text_color: Some(theme.palette().danger),
-                                    ..Default::default()
-                                }
-                            })
-                        }),
-                )
-                .push_maybe(
-                    self.upower
-                        .as_ref()
-                        .and_then(|p| p.power_profile.indicator()),
-                )
-                .push_maybe(self.audio.as_ref().and_then(|a| a.sink_indicator()))
-                .push(
-                    Row::new()
-                        .push_maybe(
-                            self.network
-                                .as_ref()
-                                .and_then(|n| n.get_connection_indicator()),
-                        )
-                        .push_maybe(self.network.as_ref().and_then(|n| n.get_vpn_indicator()))
-                        .spacing(4),
-                )
-                .push_maybe(
-                    self.upower
-                        .as_ref()
-                        .and_then(|upower| upower.battery)
-                        .map(|battery| battery.indicator()),
-                )
-                .spacing(8)
-                .into(),
-            Some(OnModulePress::ToggleMenu(MenuType::Settings)),
-        ))
-    }
-}
-
-struct AudioEventForwarder {
-    sender: ModuleEventSender<Message>,
-}
-
-impl AudioEventForwarder {
-    fn new(sender: ModuleEventSender<Message>) -> Self {
-        Self { sender }
-    }
-}
-
-impl ServiceEventPublisher<AudioService> for AudioEventForwarder {
-    type SendFuture<'a>
-        = Ready<()>
-    where
-        Self: 'a;
-
-    fn send(&mut self, event: ServiceEvent<AudioService>) -> Self::SendFuture<'_> {
-        if let Err(err) = self
-            .sender
-            .try_send(Message::Audio(AudioMessage::Event(event)))
-        {
-            warn!("failed to publish audio event: {err}");
-        }
-
-        ready(())
-    }
-}
-
-struct BrightnessEventForwarder {
-    sender: ModuleEventSender<Message>,
-}
-
-impl BrightnessEventForwarder {
-    fn new(sender: ModuleEventSender<Message>) -> Self {
-        Self { sender }
-    }
-}
-
-impl ServiceEventPublisher<BrightnessService> for BrightnessEventForwarder {
-    type SendFuture<'a>
-        = Ready<()>
-    where
-        Self: 'a;
-
-    fn send(&mut self, event: ServiceEvent<BrightnessService>) -> Self::SendFuture<'_> {
-        if let Err(err) = self
-            .sender
-            .try_send(Message::Brightness(BrightnessMessage::Event(event)))
-        {
-            warn!("failed to publish brightness event: {err}");
-        }
-
-        ready(())
-    }
-}
-
-struct NetworkEventForwarder {
-    sender: ModuleEventSender<Message>,
-}
-
-impl NetworkEventForwarder {
-    fn new(sender: ModuleEventSender<Message>) -> Self {
-        Self { sender }
-    }
-}
-
-impl ServiceEventPublisher<NetworkService> for NetworkEventForwarder {
-    type SendFuture<'a>
-        = Ready<()>
-    where
-        Self: 'a;
-
-    fn send(&mut self, event: ServiceEvent<NetworkService>) -> Self::SendFuture<'_> {
-        if let Err(err) = self
-            .sender
-            .try_send(Message::Network(NetworkMessage::Event(event)))
-        {
-            warn!("failed to publish network event: {err}");
-        }
-
-        ready(())
-    }
-}
-
-struct BluetoothEventForwarder {
-    sender: ModuleEventSender<Message>,
-}
-
-impl BluetoothEventForwarder {
-    fn new(sender: ModuleEventSender<Message>) -> Self {
-        Self { sender }
-    }
-}
-
-impl ServiceEventPublisher<BluetoothService> for BluetoothEventForwarder {
-    type SendFuture<'a>
-        = Ready<()>
-    where
-        Self: 'a;
-
-    fn send(&mut self, event: ServiceEvent<BluetoothService>) -> Self::SendFuture<'_> {
-        if let Err(err) = self
-            .sender
-            .try_send(Message::Bluetooth(BluetoothMessage::Event(event)))
-        {
-            warn!("failed to publish bluetooth event: {err}");
-        }
-
-        ready(())
-    }
-}
-
-struct UPowerEventForwarder {
-    sender: ModuleEventSender<Message>,
-}
-
-impl UPowerEventForwarder {
-    fn new(sender: ModuleEventSender<Message>) -> Self {
-        Self { sender }
-    }
-}
-
-impl ServiceEventPublisher<UPowerService> for UPowerEventForwarder {
-    type SendFuture<'a>
-        = Ready<()>
-    where
-        Self: 'a;
-
-    fn send(&mut self, event: ServiceEvent<UPowerService>) -> Self::SendFuture<'_> {
-        if let Err(err) = self
-            .sender
-            .try_send(Message::UPower(UPowerMessage::Event(event)))
-        {
-            warn!("failed to publish upower event: {err}");
-        }
-
-        ready(())
-    }
-}
-
-fn quick_settings_section<'a>(
-    buttons: Vec<(Element<'a, Message>, Option<Element<'a, Message>>)>,
-    opacity: f32,
-) -> Element<'a, Message> {
-    let mut section = column!().spacing(8);
-
-    let mut before: Option<(Element<'a, Message>, Option<Element<'a, Message>>)> = None;
-
-    for (button, menu) in buttons.into_iter() {
-        match before.take() {
-            Some((before_button, before_menu)) => {
-                section = section.push(row![before_button, button].width(Length::Fill).spacing(8));
-
-                if let Some(menu) = before_menu {
-                    section = section.push(sub_menu_wrapper(menu, opacity));
-                }
-
-                if let Some(menu) = menu {
-                    section = section.push(sub_menu_wrapper(menu, opacity));
-                }
-            }
-            _ => {
-                before = Some((button, menu));
-            }
-        }
-    }
-
-    if let Some((before_button, before_menu)) = before.take() {
-        section = section.push(
-            row![before_button, horizontal_space()]
-                .width(Length::Fill)
-                .spacing(8),
-        );
-
-        if let Some(menu) = before_menu {
-            section = section.push(sub_menu_wrapper(menu, opacity));
-        }
-    }
-
-    section.into()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{ModuleContext, event_bus::EventBus};
-    use futures::future;
-    use std::{
-        num::NonZeroUsize,
-        sync::{
-            Arc,
-            atomic::{AtomicBool, Ordering},
-        },
-    };
-    use tokio::runtime::Runtime;
-
-    #[test]
-    fn register_spawns_service_tasks() {
-        let runtime = Runtime::new().expect("runtime");
-        let bus = EventBus::new(NonZeroUsize::new(4).expect("capacity"));
-        let ctx = ModuleContext::new(bus.sender(), runtime.handle().clone());
-        let mut settings = Settings::default();
-
-        settings
-            .register(&ctx, ())
-            .expect("register should succeed");
-
-        assert!(settings.sender.is_some());
-        assert!(settings.runtime.is_some());
-        assert_eq!(settings.tasks.len(), 5);
-
-        for task in settings.tasks.drain(..) {
-            task.abort();
-        }
-    }
-
-    #[test]
-    fn register_aborts_existing_tasks() {
-        let runtime = Runtime::new().expect("runtime");
-        let bus = EventBus::new(NonZeroUsize::new(4).expect("capacity"));
-        let ctx = ModuleContext::new(bus.sender(), runtime.handle().clone());
-        let mut settings = Settings::default();
-
-        let cancelled = Arc::new(AtomicBool::new(false));
-        let guard_flag = Arc::clone(&cancelled);
-
-        settings.tasks.push(runtime.spawn(async move {
-            struct CancelGuard(Arc<AtomicBool>);
-
-            impl Drop for CancelGuard {
-                fn drop(&mut self) {
-                    self.0.store(true, Ordering::SeqCst);
-                }
-            }
-
-            let _guard = CancelGuard(guard_flag);
-
-            future::pending::<()>().await;
-        }));
-
-        settings
-            .register(&ctx, ())
-            .expect("register should succeed");
-
-        assert!(cancelled.load(Ordering::SeqCst));
-
-        for task in settings.tasks.drain(..) {
-            task.abort();
-        }
-    }
-}
-
-fn sub_menu_wrapper<Msg: 'static>(content: Element<Msg>, opacity: f32) -> Element<Msg> {
-    container(content)
-        .style(move |theme: &Theme| container::Style {
-            background: Background::Color(
-                theme
-                    .extended_palette()
-                    .secondary
-                    .strong
-                    .color
-                    .scale_alpha(opacity),
-            )
-            .into(),
-            border: Border::default().rounded(16),
-            ..container::Style::default()
-        })
-        .padding(16)
-        .width(Length::Fill)
-        .into()
-}
-
-fn quick_setting_button<'a, Msg: Clone + 'static>(
-    icon_type: Icons,
-    title: String,
-    subtitle: Option<String>,
-    active: bool,
-    on_press: Msg,
-    with_submenu: Option<(SubMenu, Option<SubMenu>, Msg)>,
-    opacity: f32,
-) -> Element<'a, Msg> {
-    let main_content = row!(
-        icon(icon_type).size(20),
-        Column::new()
-            .push(text(title).size(12))
-            .push_maybe(subtitle.map(|s| text(s).size(10)))
-            .spacing(4)
-    )
-    .spacing(8)
-    .padding(Padding::ZERO.left(4))
-    .width(Length::Fill)
-    .align_y(Alignment::Center);
-
-    button(
-        Row::new()
-            .push(main_content)
-            .push_maybe(with_submenu.map(|(menu_type, submenu, msg)| {
-                button(
-                    container(icon(if Some(menu_type) == submenu {
-                        Icons::Close
-                    } else {
-                        Icons::RightChevron
-                    }))
-                    .align_y(Vertical::Center)
-                    .align_x(Horizontal::Center),
-                )
-                .padding([4, if Some(menu_type) == submenu { 9 } else { 12 }])
-                .style(quick_settings_submenu_button_style(active, opacity))
-                .width(Length::Shrink)
-                .height(Length::Shrink)
-                .on_press(msg)
-            }))
-            .spacing(4)
-            .align_y(Alignment::Center)
-            .height(Length::Fill),
-    )
-    .padding([4, 8])
-    .on_press(on_press)
-    .height(Length::Fill)
-    .width(Length::Fill)
-    .style(quick_settings_button_style(active, opacity))
-    .width(Length::Fill)
-    .height(Length::Fixed(50.))
-    .into()
-}
+pub use audio::AudioMessage;
+pub use bluetooth::BluetoothMessage;
+pub use brightness::BrightnessMessage;
+pub use network::NetworkMessage;
+pub use power::PowerMessage;
+pub use state::{Message, Settings, SubMenu};
+pub use upower::UPowerMessage;
+pub use view::quick_setting_button;
